added all tests in one class
